+++ sypark9646/src/main/java/com/example/sypark9646/item05/Animal.java
@@ -1,4 +1,4 @@
-package com.example.sypark9646.item5;
+package com.example.sypark9646.item05;
 
 public interface Animal {
