+++ sypark9646/src/main/java/com/example/sypark9646/item05/Dog.java
@@ -1,4 +1,4 @@
-package com.example.sypark9646.item5;
+package com.example.sypark9646.item05;
 
 public class Dog implements Animal {
