+++ sypark9646/src/main/java/com/example/sypark9646/item05/Person.java
@@ -1,4 +1,4 @@
-package com.example.sypark9646.item5;
+package com.example.sypark9646.item05;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
