+++ sypark9646/src/test/java/com/example/sypark9646/item06/InstanceCreationTest.java
@@ -1,4 +1,4 @@
-package com.example.sypark9646.item6;
+package com.example.sypark9646.item06;
 
 import java.util.HashMap;
 import java.util.Map;
